split out ParserSelector and WordStats as components
